+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 # 5.1
 
+## New features
+
+- Display remaining lifes count in death screen (bleudev [#95](https://github.com/bleudev/nine_lifes/pull/95))
+
+## Fixes
+
 - Delete `EntityLifecycleEvents` as deprecation from 4.1 (sorry I forgot) (bleudev [#94](https://github.com/bleudev/nine_lifes/pull/94))
+
+### 26.1-26.2 only
+- Fixed typo in Russian localisation (bleudev [#95](https://github.com/bleudev/nine_lifes/pull/95))